Add query metrics to decision logs
Fixes #1033

+++ metrics/metrics.go
@@ -238,5 +238,5 @@ func (c *counter) Incr() {
 }
 
 func (c *counter) Value() interface{} {
-	return *c
+	return uint64(*c)
 }
